Disable cloud-init after first boot
While it does nothing on subsequent boots, if we run
cloud-init status --wait
we'll block that particular boot until it finishes concluding that is
must do nothing.
That might be a problem when a modern distro is initialized via a script
or automation tool (i.e. not via the OOBE script)
and then the user starts the instance, already fully provisioned, but
the boot blocks for a while waiting on cloud-init to determine that
there is no more work to do.