cloud-init.service is not present after 24.3.
Noble is immune to this change, but later versions won't have that
service.

Docs advise that

> Any external services which are ordered after or depend on the old
cloud-init.service name can safely switch to cloud-config.target, which
should provide the same point in boot order before and after this
change.

But that target is statically enabled

~$ systemctl is-enabled cloud-config.target
static

Not a good one to check for enablement.
This chooses cloud-init-local.service, which exists before and after
that breaking change.

Ref:
https://cloudinit.readthedocs.io/en/latest/reference/breaking_changes.html#id3