Move the integration shell script to a file
Thus we can run it inside a WSL instance at devel time, not only in CI.

Plus a small fix for questing and later: chrony instead of systemd-timesyncd

To make the test case more generic I test first if the unit is enabled
and then if active.

Since the test runs against a fresh installed instance, if the unit is
not active is because of the image state, not users disabling or masking
units etc.